+++ DentalAssistantXF/DentalAssistantXF/ViewModels/PatientHistoryPageViewModel.cs
@@ -5,6 +5,7 @@ using DentalAssistantXF.Views;
 using Prism.Commands;
 using Prism.Mvvm;
 using Prism.Navigation;
+using Prism.Services;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -16,13 +17,15 @@ namespace DentalAssistantXF.ViewModels
 	public class PatientHistoryPageViewModel : BindableBase, INavigatingAware
 	{
         private readonly INavigationService _navigationService;
+        private readonly IPageDialogService _pageDialogService;
         private readonly IDatabaseService _databaseService;
         private Patient _currentPatient;
         private PatientDentalProcedure _selectedPatientDentalProcedure;
 
-        public PatientHistoryPageViewModel(INavigationService navigationService, IDatabaseService databaseService)
+        public PatientHistoryPageViewModel(INavigationService navigationService, IPageDialogService pageDialogService, IDatabaseService databaseService)
         {
             _navigationService = navigationService;
+            _pageDialogService = pageDialogService;
             _databaseService = databaseService;
 
             MessagingCenter.Subscribe<PatientHistoryPage>(this, Constants.OnPatientHistoryPageAppearingMsg, (sender) => { GetPatienDentalOperationsAsync(); }); //{ GetPatienDentalOperationsAsync(); });
@@ -49,12 +52,15 @@ namespace DentalAssistantXF.ViewModels
 
         public DelegateCommand<PatientDentalProcedure> DeleteProcedureCommand => new DelegateCommand<PatientDentalProcedure>(async (args) =>
         {
-            if (await _databaseService.DentalAssistantDB.DeleteProcedureAsync(args) > 0)
+            if (await _pageDialogService.DisplayAlertAsync("Alert", "Are you sure you want to delete this procedure?", "Yes", "Cancel"))
             {
-                PatientDentalProcedures.Remove(args);
-                HelperFunctions.ShowToastMessage(ToastMessageType.Success, "Dental procedure deleted successfully");
-                MessagingCenter.Send(this, Constants.OnDashboardDataChangeMsg);
-                MessagingCenter.Send(this, Constants.OnAddOrEditPatientMsg);
+                if (await _databaseService.DentalAssistantDB.DeleteProcedureAsync(args) > 0)
+                {
+                    GetPatienDentalOperationsAsync();
+                    HelperFunctions.ShowToastMessage(ToastMessageType.Success, "Dental procedure deleted successfully");
+                    MessagingCenter.Send(this, Constants.OnDashboardDataChangeMsg);
+                    MessagingCenter.Send(this, Constants.OnAddOrEditPatientMsg);
+                }
             }
         });
 
@@ -73,13 +79,9 @@ namespace DentalAssistantXF.ViewModels
             {
                 procedures.Last().IsLast = true;
             }
-
-            PatientDentalProcedures.Clear();
             
-            foreach(var procedure in procedures)
-            {
-                PatientDentalProcedures.Add(procedure);
-            }
+            PatientDentalProcedures.Clear();
+            procedures.ForEach(PatientDentalProcedures.Add);
         }
 
         private async void AddOrEditPatientDentalProcedureAsync(string mode = "Edit")
